+++ R/metadata.R
@@ -54,3 +54,17 @@ get_metadata <- function() {
   naomi_read_csv(system_file("metadata", "metadata.csv"))
 }
 
+#' Get 5 year age groups
+#'
+#'
+#' @return ID for 5 year age groups
+#'
+#' @export
+#'
+#' @examples
+#' get_five_year_age_groups()
+get_five_year_age_groups <- function() {
+  age_groups <- get_age_groups()
+  age_groups <- age_groups[age_groups$age_group_span == 5 | age_groups$age_group == "80+", ]
+  age_groups$age_group
+}