+++ tests/testthat/test-run-model.R
@@ -7,7 +7,7 @@ test_that("model can be run", {
     shape = system_file("extdata/areas/area_merged.geojson"),
     survey = system_file("extdata/survey/survey_hiv_indicators.csv"),
     art_number = system_file("extdata/programme/art_number.csv"),
-    anc_testing= system_file("extdata/programme/anc_testing.csv")
+    anc_testing = system_file("extdata/programme/anc_testing.csv")
   )
   options <- list(
     area_scope = "MWI",